Remove PropertyBuilder abstraction, inline into consolidate-struct-assignments
