+++ src/plugin/src/transforms/consolidate-struct-assignments.ts
@@ -5,8 +5,10 @@
 import { Core, type MutableGameMakerAstNode } from "@gml-modules/core";
 import { FunctionalParserTransform } from "./functional-transform.js";
 import { CommentTracker } from "./utils/comment-tracker.js";
-import { StructAssignmentMatcher } from "./utils/struct-assignment-matcher.js";
-import { PropertyBuilder } from "./utils/property-builder.js";
+import {
+    StructAssignmentMatcher,
+    type AssignmentDetails
+} from "./utils/struct-assignment-matcher.js";
 import { AssignmentCommentHandler } from "./utils/assignment-comment-handler.js";
 
 type CommentTools = {
@@ -20,6 +22,10 @@ type ConsolidateStructAssignmentsTransformOptions = {
 // Avoid destructuring Core across package boundaries; call Core.* functions
 // directly. This prevents assumptions about nested namespaces and matches
 // the monorepo conventions in AGENTS.md.
+
+const ASSIGNMENT_EXPRESSION = "AssignmentExpression";
+const IDENTIFIER = "Identifier";
+const LITERAL = "Literal";
 
 const FALLBACK_COMMENT_TOOLS = Object.freeze({
     addTrailingComment() {}
@@ -79,17 +85,131 @@ function restoreComment(comment, snapshot) {
     }
 }
 
+type PropertyKeyInfo = {
+    identifierName: string;
+    raw: unknown;
+    start: unknown;
+    end: unknown;
+};
+
+function getPropertyKeyInfo(propertyNode: unknown): PropertyKeyInfo | null {
+    if (!Core.isNode(propertyNode)) {
+        return null;
+    }
+
+    if (Core.isIdentifierNode(propertyNode)) {
+        return {
+            identifierName: propertyNode.name,
+            raw: propertyNode.name,
+            start: propertyNode.start,
+            end: propertyNode.end
+        };
+    }
+
+    if (
+        Core.isLiteralNode(propertyNode) &&
+        typeof propertyNode.value === "string"
+    ) {
+        const unquoted = Core.stripStringQuotes(propertyNode.value as any);
+        return {
+            identifierName: unquoted,
+            raw: propertyNode.value,
+            start: propertyNode.start,
+            end: propertyNode.end
+        };
+    }
+
+    return null;
+}
+
+function buildPropertyNameNode(
+    propertyKey: PropertyKeyInfo | null,
+    isIdentifierSafe: (name: unknown) => boolean
+): unknown {
+    if (!propertyKey) {
+        return null;
+    }
+
+    const identifierName = propertyKey.identifierName;
+    if (identifierName && isIdentifierSafe(identifierName)) {
+        return {
+            type: IDENTIFIER,
+            name: identifierName,
+            start: Core.cloneLocation(propertyKey.start) ?? null,
+            end: Core.cloneLocation(propertyKey.end) ?? null
+        };
+    }
+
+    if (typeof propertyKey.raw === "string") {
+        return {
+            type: LITERAL,
+            value: propertyKey.raw,
+            start: Core.cloneLocation(propertyKey.start) ?? null,
+            end: Core.cloneLocation(propertyKey.end) ?? null
+        };
+    }
+
+    return null;
+}
+
+function getPreferredLocation(primary: unknown, fallback: unknown): unknown {
+    if (Core.isNode(primary)) {
+        return primary;
+    }
+    if (Core.isNode(fallback)) {
+        return fallback;
+    }
+    return null;
+}
+
+function buildPropertyFromAssignment(
+    assignmentDetails: AssignmentDetails | null,
+    isIdentifierSafe: (name: unknown) => boolean
+): MutableGameMakerAstNode | null {
+    if (!assignmentDetails) {
+        return null;
+    }
+
+    const { assignment, propertyAccess } = assignmentDetails;
+    if (!Core.isNode(assignment) || assignment.type !== ASSIGNMENT_EXPRESSION) {
+        return null;
+    }
+
+    if (!propertyAccess) {
+        return null;
+    }
+
+    const propertyKey = getPropertyKeyInfo(propertyAccess.propertyNode);
+    const propertyName = buildPropertyNameNode(propertyKey, isIdentifierSafe);
+    if (!propertyName) {
+        return null;
+    }
+
+    return {
+        type: "Property",
+        name: propertyName,
+        value: assignment.right,
+        start:
+            Core.cloneLocation(
+                getPreferredLocation(
+                    propertyAccess.propertyStart,
+                    assignment.start
+                )
+            ) ?? null,
+        end:
+            Core.cloneLocation(
+                getPreferredLocation(assignment.right?.end, assignment.end)
+            ) ?? null
+    } as unknown as MutableGameMakerAstNode;
+}
+
 export class ConsolidateStructAssignmentsTransform extends FunctionalParserTransform<ConsolidateStructAssignmentsTransformOptions> {
     private readonly matcher: StructAssignmentMatcher;
-    private readonly propertyBuilder: PropertyBuilder;
     private readonly commentHandler: AssignmentCommentHandler;
 
     constructor() {
         super("consolidate-struct-assignments", {});
         this.matcher = new StructAssignmentMatcher();
-        this.propertyBuilder = new PropertyBuilder(
-            this.matcher.isIdentifierSafe.bind(this.matcher)
-        );
         this.commentHandler = new AssignmentCommentHandler();
     }
 
@@ -192,10 +312,8 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
             // The collected properties are assigned to the struct node; cast to
             // `MutableGameMakerAstNode` to allow mutation in-place with correct
             // typing for downstream transforms.
-            (structNode).properties =
-                collected.properties;
-            (structNode).hasTrailingComma =
-                collected.shouldForceBreak;
+            structNode.properties = collected.properties;
+            structNode.hasTrailingComma = collected.shouldForceBreak;
 
             statements.splice(index + 1, collected.count);
         }
@@ -251,10 +369,10 @@ export class ConsolidateStructAssignmentsTransform extends FunctionalParserTrans
                 break;
             }
 
-            const property =
-                this.propertyBuilder.buildPropertyFromAssignment(
-                    assignmentDetails
-                );
+            const property = buildPropertyFromAssignment(
+                assignmentDetails,
+                this.matcher.isIdentifierSafe.bind(this.matcher)
+            );
             if (!property) {
                 break;
             }
